requirements.txt
pycryptodome
cryptography
colorama
